Fixed navigation space issue & moved code to another file
