+++ Splito/UI/Home/Groups/Group/More Options/Transaction List/TransactionListView.swift
@@ -1,0 +1,18 @@
+//
+//  TransactionListView.swift
+//  Splito
+//
+//  Created by Nirali Sonani on 14/06/24.
+//
+
+import SwiftUI
+
+struct TransactionListView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    TransactionListView()
+}
