+++ Splito/UI/Home/Groups/Group/More Options/Transaction List/TransactionListViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  TransactionListViewModel.swift
+//  Splito
+//
+//  Created by Nirali Sonani on 14/06/24.
+//
+
+import Foundation
